requirements.txt
alembic==1.9.4
anyio==3.6.2
boto3==1.26.102
botocore==1.29.102
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
loguru==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2==2.9.5
pydantic==1.10.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
win32-setctime==1.1.0
